refactor(chat): extract message sync into reusable component
Move the fetch-and-persist core (getAndPersistMessages,
persistChatMessagesAndHandleSystemMessages, updateBlocks) out of
OfflineFirstChatRepository into a singleton ChatMessageSyncer that
takes (user, roomToken, threadId) per call instead of relying on
lateinit state set by ChatActivity. The repository delegates to it
so open-chat and background paths share a single write path.
